third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/service-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHealthThermometerDevice({
      filters: [{services: ['health_thermometer']}],
      optionalServices: ['generic_access']})
    .then(({device}) => {
      return Promise.all([
          device.gatt.getPrimaryService(generic_access.alias),
          device.gatt.getPrimaryService(generic_access.name),
          device.gatt.getPrimaryService(generic_access.uuid)])
        .then(services => {
          services.forEach(service => {
            assert_equals(service.uuid, generic_access.uuid,
                          'Service UUID should be the same as requested UUID.');
            assert_true(service.isPrimary,
                        'getPrimaryService should return a primary service.');
            assert_equals(service.device, device,
                          'Service device should be the same as device.');
          });
        });
    });
}, 'Request for service. Should return right service');
</script>
